move product crud to container
